organize buffer related classes in their own sub-package, move subclasses out of interface
